requirements.txt
# ABM Research System Dependencies
requests>=2.31.0
openai>=1.3.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
pandas>=2.1.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
httpx>=0.25.0
aiohttp>=3.8.0
pathlib2>=2.3.0
python-dateutil>=2.8.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
